Move ClipID to clip_id.hpp/cpp
