+++ src/mbgl/renderer/painter_clipping.cpp
@@ -3,7 +3,7 @@
 #include <mbgl/map/map.hpp>
 #include <mbgl/map/tile.hpp>
 #include <mbgl/map/source.hpp>
-#include <mbgl/util/clip_ids.hpp>
+#include <mbgl/util/clip_id.hpp>
 
 using namespace mbgl;
